89831: SystemML compiler/runtime issues - Fix determine number of reducers mr reblock (integer overflow number of reducer groups if more than 2g blocks)

+++ SystemML/SystemML/src/main/java/com/ibm/bi/dml/runtime/matrix/CSVReblockMR.java
@@ -354,7 +354,7 @@ public class CSVReblockMR
 		MatrixCharacteristics[] stats=ret.stats;
 		
 		//set up the number of reducers
-		int numRed = WriteCSVMR.determineNumReducers(rlens, clens, ConfigurationManager.getConfig().getIntValue(DMLConfig.NUM_REDUCERS), (int)ret.numReducerGroups);
+		int numRed = WriteCSVMR.determineNumReducers(rlens, clens, ConfigurationManager.getConfig().getIntValue(DMLConfig.NUM_REDUCERS), ret.numReducerGroups);
 		job.setNumReduceTasks( numRed );
 		
 		// Print the complete instruction
