+++ SystemML/SystemML/src/main/java/com/ibm/bi/dml/runtime/matrix/WriteCSVMR.java
@@ -152,7 +152,7 @@ public class WriteCSVMR
 	 * @param numRedGroups
 	 * @return
 	 */
-	protected static int determineNumReducers( long[] rlen, long[] clen, int defaultNumRed, int numRedGroups )
+	protected static int determineNumReducers( long[] rlen, long[] clen, int defaultNumRed, long numRedGroups )
 	{
 		//init return with default value
 		int ret = defaultNumRed;
@@ -174,7 +174,7 @@ public class WriteCSVMR
 		ret = (int)Math.max(ret, Math.min(maxSize/blockSize, maxNumRed));
 		
 		//reduce num reducers for few result blocks
-		ret = Math.min(ret, numRedGroups);
+		ret = (int)Math.min(ret, numRedGroups);
 		
 		//ensure there is at least one reducer
 		ret = Math.max(ret, 1);
